cleaned up store packages, removed height cache
