--- store/rootmulti/dbadapter.go
@@ -1,31 +1,0 @@
-package rootmulti
-
-import (
-	"github.com/pokt-network/pocket-core/store/dbadapter"
-	"github.com/pokt-network/pocket-core/store/types"
-)
-
-var commithash = []byte("FAKE_HASH")
-
-//----------------------------------------
-// commitDBStoreWrapper should only be used for simulation/debugging,
-// as it doesn't compute any commit hash, and it cannot load older state.
-
-// Wrapper type for dbm.Db with implementation of KVStore
-type commitDBStoreAdapter struct {
-	dbadapter.Store
-}
-
-func (cdsa commitDBStoreAdapter) Commit() types.CommitID {
-	return types.CommitID{
-		Version: -1,
-		Hash:    commithash,
-	}
-}
-
-func (cdsa commitDBStoreAdapter) LastCommitID() types.CommitID {
-	return types.CommitID{
-		Version: -1,
-		Hash:    commithash,
-	}
-}
