+++ store/rootmulti/store_test.go
@@ -14,7 +14,7 @@ import (
 
 func TestStoreType(t *testing.T) {
 	db := dbm.NewMemDB()
-	store := NewStore(db, false, 5000000)
+	store := NewStore(db)
 	store.MountStoreWithDB(
 		types.NewKVStoreKey("store1"), types.StoreTypeIAVL, db)
 
@@ -22,7 +22,7 @@ func TestStoreType(t *testing.T) {
 
 func TestStoreMount(t *testing.T) {
 	db := dbm.NewMemDB()
-	store := NewStore(db, false, 5000000)
+	store := NewStore(db)
 
 	key1 := types.NewKVStoreKey("store1")
 	key2 := types.NewKVStoreKey("store2")
@@ -33,38 +33,6 @@ func TestStoreMount(t *testing.T) {
 
 	require.Panics(t, func() { store.MountStoreWithDB(key1, types.StoreTypeIAVL, db) })
 	require.Panics(t, func() { store.MountStoreWithDB(dup1, types.StoreTypeIAVL, db) })
-}
-
-func TestCacheMultiStoreWithVersion(t *testing.T) {
-	var db dbm.DB = dbm.NewMemDB()
-	ms := newMultiStoreWithMounts(db)
-	err := ms.LoadLatestVersion()
-	require.Nil(t, err)
-
-	commitID := types.CommitID{}
-	checkStore(t, ms, commitID, commitID)
-
-	k, v := []byte("wind"), []byte("blows")
-
-	store1 := ms.getStoreByName("store1").(types.KVStore)
-	_ = store1.Set(k, v)
-
-	cID := ms.Commit()
-	require.Equal(t, int64(1), cID.Version)
-
-	// require failure when given an invalid or pruned version
-	_, err = ms.CacheMultiStoreWithVersion(cID.Version + 1)
-	require.Error(t, err)
-
-	// require a valid version can be cache-loaded
-	cms, err := ms.CacheMultiStoreWithVersion(cID.Version)
-	require.NoError(t, err)
-
-	// require a valid Key lookup yields the correct value
-	kvStore := cms.GetKVStore(ms.keysByName["store1"])
-	require.NotNil(t, kvStore)
-	kg, _ := kvStore.Get(k)
-	require.Equal(t, kg, v)
 }
 
 func TestHashStableWithEmptyCommit(t *testing.T) {
@@ -246,7 +214,7 @@ func TestMultiStoreQuery(t *testing.T) {
 // utils
 
 func newMultiStoreWithMounts(db dbm.DB) *Store {
-	store := NewStore(db, false, 5000000)
+	store := NewStore(db)
 	store.MountStoreWithDB(
 		types.NewKVStoreKey("store1"), types.StoreTypeIAVL, nil)
 	store.MountStoreWithDB(
